sfc: Fix polling for slow MCDI operations
https://lists.ubuntu.com/archives/kernel-team/2010-January/008303.html

When the interface is down and we are using polled mode for MCDI
operations, we busy-wait for completion for approximately 1 jiffy
using udelay() and then back off to schedule().  But the completion
will not wake the task, since we are using polled mode!  We must use
schedule_timeout_uninterruptible() instead.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 55029c1d65158aea9672c5dfadb43a57f23e3100)

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>