UBUNTU: [Upstream] e1000: enhance frame fragment detection
CVE-2009-4536

OriginalAuthor: Neil Horman <nhorman@redhat.com>
OriginalLocation: http://marc.info/?l=linux-netdev&m=126203101730472&w=2

A security discussion was recently given:
http://events.ccc.de/congress/2009/Fahrplan//events/3596.en.html
And a patch that was submitted awhile back was brought up.  Apparently
some of their testing revealed that they were able to force a buffer
fragment in e1000 in which the trailing fragment was greater than 4
bytes.  As a result the fragment check introduced failed to detect the
fragement and a partial invalid frame was passed up into the network
stack.  This patch corrects it.  Neil is in the process of testing it
now, but it makes good logical sense to him.  Effectively it maintains a
per-adapter state variable which detects a non-EOP frame, and discards
it and subsequent non-EOP frames leading up to _and_ _including_ the
next positive-EOP frame (as it is by definition the last fragment).
This should prevent any and all partial frames from entering the network
stack from e1000

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>